get user id from jwt

+++ AspAPI/ExamBanking/ExamBanking/Models/Payment.cs
@@ -6,7 +6,7 @@ namespace ExamBanking.Models
     public partial class Payment
     {
         public int Payid { get; set; }
-        public int? Accid { get; set; }
+        public decimal? Accid { get; set; }
         public DateTime? Paydate { get; set; }
         public int? Paymode { get; set; }
         public int? Money { get; set; }
